Conflicts:
	code/qcommon/files.cpp
	code/qcommon/files_common.cpp
	code/qcommon/qcommon.h

+++ code/ui/ui_public.h
@@ -55,7 +55,7 @@ typedef struct {
 	int 		(*FS_Write)( const void *buffer, int len, fileHandle_t f );
 	void		(*FS_FCloseFile)( fileHandle_t f );
 	int			(*FS_GetFileList)(  const char *path, const char *extension, char *listbuf, int bufsize );
-	int			(*FS_ReadFile)( const char *name, void **buf );
+	long		(*FS_ReadFile)( const char *name, void **buf );
 	void		(*FS_FreeFile)( void *buf );
 
 	// =========== renderer function calls ================
